Velocity Cacher and Wall Stop
Wall Stop - Override CharacterBody Velocity when hitting a wall. Otherwise, the Velocity would remain the same, and if the player reaches the edge of the wall before losing its speed, he will be launched.
Velocity Cacher - This Edge launch movement is a feature we want, but we do not want to happen every time. We added a Velocity Cache, which is filled when entering a wall, and used when leaving a wall.
By default, it has a 100ms lifespan.