Crouch Handling
This is a big dirty commit resulting of quite a lot of work, here is a breakdown.
The crouch input can result in various behaviours, to deal with this, we built a specific architecture.

PI_CrouchDispatcher dispatches the KeyUp and KeyDown inputs to the Dash and SlideInput handlers.

PI_Slide tries to consume the incoming input (up or down), and propagates to PI_Crouch if it can't.
PI_Dash is an end consumer, so it will just consume. Note that Slide will always consume in parrallel of PI_Slide.

PI_Crouch is the end consumer of the PI_Slide branch.

PI_Sprint gets cancelled by PI_Crouch and PI_Slide.

PI_Holdable is an interface built to help deal with configurable inputs on hold/press.
Not used in this version of PI_CrouchDispatcher, but might be usefull for many other things.