PI_CrouchDispatcher - Improved and clearified responsibilities.
PI_Crouch and PI_Slide do not need any reference to the dispatcher anymore.
The Dispatcher is now in charge of handling hold/tap mode, which was already implictly the case as crouch and slide tap mode should be bounded, and were acting bounded.
Added PHX_Checks for the later addition of shrinkability checks in the Dispatcher.