Camera deformation fix + investigate interpollation warning
Warning is benign on camera sprites. It happens because godot considers the position of the remote position is updated on physics process, although it's interpolated on process. Then, it sees the sprites have been moved on physics process, but really, they are on process since the remote position do update the interpolated positions