PW_Weapon - Subtick firing
There were a kind of bug with weapons that had high fire rate. Indeed, I use a timer to trigger the next shot, but the event might be triggered one tick late if the timing doesn't align perfectly. One single tick can represent a big difference if the fire rate is near the tick rate, and drastically reduce dps. I thus implemented two method, but one might be irrelevant.

The revelant for sure, is to compute the difference in time between what delay we expected between two shots, and what delay we actually got. Then, we deduce that unaccuracy from next shot expected time.

The other solution, that I think is less interesting, is to use partial shots. That is, a shot that fires at the same time as the initial shot, but with partial power, so partial damage, knockback, recoil, etc. We can shoot 1,2 shots in a tick for example. It's probably less relevant as it is less accurate to what it would actually look like it if was continuous, for example hit sound will still be too slow, and damages might feel off.